+++ src/components/ui/toolbox.ts
@@ -106,10 +106,10 @@ export default class Toolbox extends EventsDispatcher<ToolboxEventMap> {
 
   /**
    * CSS styles
-   *
-   * @returns {Object<string, string>}
    */
-  private static get CSS(): { [name: string]: string } {
+  private static get CSS(): {
+    toolbox: string;
+    } {
     return {
       toolbox: 'ce-toolbox',
     };
@@ -153,6 +153,10 @@ export default class Toolbox extends EventsDispatcher<ToolboxEventMap> {
 
     this.nodes.toolbox = this.popover.getElement();
     this.nodes.toolbox.classList.add(Toolbox.CSS.toolbox);
+
+    if (import.meta.env.MODE === 'test') {
+      this.nodes.toolbox.setAttribute('data-cy', 'toolbox');
+    }
 
     return this.nodes.toolbox;
   }