+++ test/cypress/tests/utils/flipper.cy.ts
@@ -38,7 +38,6 @@ class SomePlugin {
 
 describe('Flipper', () => {
   it('should prevent plugins event handlers from being called while keyboard navigation', () => {
-    const TAB_KEY_CODE = 9;
     const ARROW_DOWN_KEY_CODE = 40;
     const ENTER_KEY_CODE = 13;
 
@@ -63,14 +62,16 @@ describe('Flipper', () => {
 
     cy.get('[data-cy=editorjs]')
       .get('.cdx-some-plugin')
+      .as('pluginInput')
       .focus()
-      .type(sampleText);
+      .type(sampleText)
+      .wait(100);
 
     // Try to delete the block via keyboard
     cy.get('[data-cy=editorjs]')
       .get('.cdx-some-plugin')
       // Open tunes menu
-      .trigger('keydown', { keyCode: TAB_KEY_CODE })
+      .trigger('keydown', { code: 'Slash', ctrlKey: true })
       // Navigate to delete button (the second button)
       .trigger('keydown', { keyCode: ARROW_DOWN_KEY_CODE })
       .trigger('keydown', { keyCode: ARROW_DOWN_KEY_CODE });